fix meta.tables, allow rest endpoint to use http, move tests,
