+++ resources/config/nginx.conf
@@ -60,8 +60,23 @@ http {
         listen 80 default_server;
         listen [::]:80 default_server;
 
-        # Redirect all HTTP requests to HTTPS with a 301 Moved Permanently response.
-        return 301 https://$host$request_uri;
+        server_name         activedata.allizom.org;
+
+        location / {
+            # Redirect all HTTP requests to HTTPS with a 301 Moved Permanently response.
+            return 301 https://$host$request_uri;
+        }
+
+        location /tuid {
+            proxy_pass http://tuid;
+            proxy_read_timeout 300;
+        }
+
+        location /query {
+            proxy_pass http://backend;
+            proxy_read_timeout 300;
+        }
+
     }
 
     server {
@@ -89,7 +104,7 @@ http {
             proxy_read_timeout 300;
         }
 
-            location /tuid {
+        location /tuid {
             proxy_pass http://tuid;
             proxy_read_timeout 300;
         }
