+++ resources/config/nginx_testing.conf
@@ -58,6 +58,8 @@ http {
 
     server {
         listen              80 default_server;
+        listen              [::]:80 default_server;
+
         server_name         activedata.allizom.org;
 
         location / {
